# frontend/src/openapi/openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8090/api/v1","description":"Generated server url"}],"paths":{"/messages":{"post":{"tags":["Message"],"operationId":"saveMessage","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageRequest"}}},"required":true},"responses":{"201":{"description":"Created"}}},"patch":{"tags":["Message"],"operationId":"setMessagesToSeen","parameters":[{"name":"chat-id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"202":{"description":"Accepted"}}}},"/messages/upload-file":{"post":{"tags":["Message"],"operationId":"uploadMedia","parameters":[{"name":"chat-id","in":"query","required":true,"schema":{"type":"string"}},{"name":"media-type","in":"query","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"multipart/form-data":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"201":{"description":"Created"}}}},"/chats":{"get":{"tags":["Chat"],"operationId":"getChatsByReceiver","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ChatResponse"}}}}}}},"post":{"tags":["Chat"],"operationId":"createChat","parameters":[{"name":"sender-id","in":"query","required":true,"schema":{"type":"string"}},{"name":"receiver-id","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StringResponse"}}}}}}},"/chats/{chatId}/favorite":{"get":{"tags":["Chat"],"operationId":"isFavoriteForUser","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"boolean"}}}}}},"post":{"tags":["Chat"],"operationId":"addToFavorite","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}},"delete":{"tags":["Chat"],"operationId":"removeFromFavorite","parameters":[{"name":"chatId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"}}}},"/users":{"get":{"tags":["User"],"operationId":"getAllUsers","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/UserResponse"}}}}}}}},"/messages/chat/{chat-id}":{"get":{"tags":["Message"],"operationId":"getMessages","parameters":[{"name":"chat-id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/MessageResponse"}}}}}}}}},"components":{"schemas":{"MessageRequest":{"type":"object","properties":{"content":{"type":"string"},"senderId":{"type":"string"},"receiverId":{"type":"string"},"type":{"type":"string","enum":["TEXT","IMAGE","AUDIO","VIDEO"]},"chatId":{"type":"string"}}},"StringResponse":{"type":"object","properties":{"response":{"type":"string"}}},"UserResponse":{"type":"object","properties":{"id":{"type":"string"},"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"lastSeen":{"type":"string","format":"date-time"},"online":{"type":"boolean"}}},"MessageResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"content":{"type":"string"},"type":{"type":"string","enum":["TEXT","IMAGE","AUDIO","VIDEO"]},"state":{"type":"string","enum":["SENT","SEEN"]},"senderId":{"type":"string"},"receiverId":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"media":{"type":"array","items":{"type":"string","format":"byte"}}}},"ChatResponse":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"unreadCount":{"type":"integer","format":"int64"},"lastMessage":{"type":"string"},"lastMessageTime":{"type":"string","format":"date-time"},"senderId":{"type":"string"},"receiverId":{"type":"string"},"recipientOnline":{"type":"boolean"},"favorite":{"type":"boolean"}}}},"securitySchemes":{"keycloak":{"type":"oauth2","in":"header","scheme":"bearer","bearerFormat":"JWT","flows":{"password":{"authorizationUrl":"http://localhost:9090/realms/echohub/protocol/openid-connect/auth","tokenUrl":"http://localhost:9090/realms/echohub/protocol/openid-connect/token","scopes":{}}}}}}}