# src/main/java/pl/dolien/echohub/message/MessageController.java
package pl.dolien.echohub.message;

import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.tags.Tag;
import lombok.RequiredArgsConstructor;
import org.springframework.security.core.Authentication;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import pl.dolien.echohub.message.dto.MessageRequest;
import pl.dolien.echohub.message.dto.MessageResponse;

import java.util.List;

import static org.springframework.http.HttpStatus.ACCEPTED;
import static org.springframework.http.HttpStatus.CREATED;

@RestController
@RequestMapping("/messages")
@RequiredArgsConstructor
@Tag(name = "Message")
public class MessageController {

    private final MessageService service;

    @PostMapping
    @ResponseStatus(CREATED)
    public void saveMessage(@RequestBody MessageRequest message) {
        service.saveMessage(message);
    }

    @PostMapping(value = "/upload-file", consumes = "multipart/form-data")
    @ResponseStatus(CREATED)
    public void uploadMedia(
            @RequestParam("chat-id") String chatId,
            @Parameter()
            @RequestParam("file") MultipartFile file,
            @RequestParam("media-type") MessageType messageType,
            Authentication auth
    ) {
        service.uploadMediaMessage(chatId, file, auth, messageType);
    }

    @PatchMapping
    @ResponseStatus(ACCEPTED)
    public void setMessagesToSeen(
            @RequestParam("chat-id") String chatId,
            Authentication auth
    ) {
        service.setMessagesToSeen(chatId, auth);
    }

    @GetMapping("/chat/{chat-id}")
    public List<MessageResponse> getMessages(
            @PathVariable("chat-id") String chatId
    ) {
        return service.getChatMessages(chatId);
    }
}